Rename tony file to the old one.
