+++ TombEngine/Objects/TR3/Entity/tr3_tony.h
