+++ TombEngine/TombEngine.vcxproj
@@ -170,7 +170,7 @@ CALL gen.bat</Command>
     <ClInclude Include="Math\Objects\AxisAngle.h" />
     <ClInclude Include="Objects\TR1\Trap\DamoclesSword.h" />
     <ClInclude Include="Objects\TR3\Entity\Compsognathus.h" />
-    <ClInclude Include="Objects\TR3\Entity\TonyBoss.h" />
+    <ClInclude Include="Objects\TR3\Entity\tr3_tony.h" />
     <ClInclude Include="Objects\TR5\Entity\HeavyGuard.h" />
     <ClInclude Include="Objects\TR3\Entity\Lizard.h" />
     <ClInclude Include="Objects\TR3\Entity\PunaBoss.h" />
@@ -649,7 +649,7 @@ CALL gen.bat</Command>
     <ClCompile Include="Math\Objects\AxisAngle.cpp" />
     <ClCompile Include="Objects\TR1\Trap\DamoclesSword.cpp" />
     <ClCompile Include="Objects\TR3\Entity\Compsognathus.cpp" />
-    <ClCompile Include="Objects\TR3\Entity\TonyBoss.cpp" />
+    <ClCompile Include="Objects\TR3\Entity\tr3_tony.cpp" />
     <ClCompile Include="Objects\TR5\Entity\HeavyGuard.cpp" />
     <ClCompile Include="Objects\TR3\Entity\Lizard.cpp" />
     <ClCompile Include="Objects\TR3\Entity\PunaBoss.cpp" />
